+++ src/test/neural_network_test.rs
@@ -10,6 +10,7 @@ mod average_pooling_3d_test;
 mod conv1d_test;
 mod conv2d_test;
 mod conv3d_test;
+mod dense_test;
 mod depthwise_conv2d_test;
 mod flatten_test;
 mod global_average_pooling_1d_test;